requirements.txt
discord.py
python-dotenv
riotwatcher
discord-pretty-help
python-dateutil
pytz
ffmpeg-python
youtube_dl
youtube-search
